added 15 tests covering RFC 6578 sync-collection for CalDAV + CardDAV: initial sync, no-op delta, delta filtering, tombstone-on-delete cross-checked against ground truth, cross-collection token rejection, churn collapsing, concurrent-writer race pin, and a row-cap/507-expiry/recovery test — plus two "local mirror built purely from deltas must exactly equal server ground truth after N rounds of churn" stress tests.   Infra: run-pycaldav.sh now exports OXICLOUD_ADMIN_JWT and sets a small OXICLOUD_SYNC_LOG_MAX_ROWS_PER_COLLECTION=5 so the row-cap/expiry path is deterministically testable; conftest.py got matching admin_jwt/base_url fixtures. Bugfixes   1. SUM(bigint) decodes as NUMERIC in Postgres, not i64 — broke the retention job's count query outright (every sweep failed). Fixed with ::bigint casts in both PG repos.   2. Stale-tombstone bug: when a href is deleted+recreated within one poll window (CalDAV/CardDAV "update" = delete+recreate, new member_id), the response renderer buckets upserts and deletions into separate lists with no   ordering between them — a client could receive a tombstone for a href that's actually still live and wrongly delete something that exists. Fixed by dropping same-batch stale tombstones server-side in sync_collection_engine.rs and webdav_sync_collection_service.rs.

+++ src/infrastructure/repositories/pg/folder_sync_change_pg_repository.rs
@@ -54,7 +54,15 @@ impl FolderSyncChangeRepository for FolderSyncChangePgRepository {
 
         // DISTINCT ON collapses churn within the window to the latest row
         // per member (e.g. trash-then-restore nets to the correct single
-        // outcome instead of contradictory duplicate entries).
+        // outcome instead of contradictory duplicate entries). Row order
+        // here doesn't matter to the client: the handler buckets upserts
+        // and deletions into separate lists before rendering (see
+        // `webdav_handler.rs`'s manual split / `split_homogeneous`), so
+        // no response ever preserves this query's row order anyway — the
+        // real hazard that shape creates (a stale tombstone can outlive
+        // a same-href recreation within one poll window) is handled by
+        // dropping such tombstones before rendering, not by row order
+        // here. See `webdav_sync_collection_service.rs`'s comment.
         let rows = sqlx::query_as::<_, (i64, String, Uuid, String, String)>(
             r#"
             SELECT DISTINCT ON (member_id)
@@ -151,7 +159,7 @@ impl FolderSyncChangeRepository for FolderSyncChangePgRepository {
                         EXCLUDED.low_water_seq
                     )
             )
-            SELECT COALESCE(SUM(n), 0) FROM per_collection
+            SELECT COALESCE(SUM(n), 0)::bigint FROM per_collection
             "#,
         )
         .bind(cutoff)
@@ -188,7 +196,7 @@ impl FolderSyncChangeRepository for FolderSyncChangePgRepository {
                         EXCLUDED.low_water_seq
                     )
             )
-            SELECT COALESCE(SUM(n), 0) FROM per_collection
+            SELECT COALESCE(SUM(n), 0)::bigint FROM per_collection
             "#,
         )
         .bind(max_rows as i64)
